Add Nyaa provider
If your looking into this commit and wondering if this will mess anything up for Verizon users, its not.

If you enabled both provider.nyaa and provider.nyaaalt settings, the _BASE_URL will be set to 'https://nyaa.iss.ink/'.

Here's why:

In the code you provided, _BASE_URL is initialized using two if statements. If provider.nyaa is true, _BASE_URL is set to 'https://nyaa.si/'. If provider.nyaaalt is true, _BASE_URL is set to 'https://nyaa.iss.ink/'.

If both provider.nyaa and provider.nyaaalt are true, the first if statement will set _BASE_URL to 'https://nyaa.si/', but then the second if statement will overwrite it with 'https://nyaa.iss.ink/'. So, the final value of _BASE_URL will be 'https://nyaa.iss.ink/'.