Update AniListBrowser.py
The Airing Calendar is a feature I really want to use when the new season hits but I just can't handle all the Chinese and Korean releases flooding the page. Since the AiringSchedules object doesn't have a filter for countryOfOrigin, I had to add it to the subsequent filtering step. Not too happy doing that filter client side but it's better than nothing.