Updated Version
Changelog 0.4.19 [20230821]:
- The Final Permanent Fix For Audio and Subtitles.

This took me half a day to figure out cause there was no kodi documentation to get audio name of the stream. Sense the only two audios were "eng" and "jpn", we could use elimination. If one source did not have "eng" or "jpn" it would activate either one or the other depending on what's missing. If there were more then one source such as dual-audio, it would go by the setting we have selected and run its part of the code that we picked by the setting.

This was overlay complicated but im surprised to have found a permanent solution to all Audio and Subtitle Issues.